Xilskey: Signed/unsigned conversion without cast.
This patch fixes below Misra-c violation:
        Signed/unsigned conversion without cast.
        MISRA-C:2012 R.10.3,R.10.4


Acked-for-series: VNSL Durga Challa <vnsl.durga.challa@xilinx.com>

+++ lib/sw_services/xilskey/src/xilskey_eps_zynqmp.c
@@ -514,7 +514,7 @@ static inline u32 XilSKey_ZynqMp_EfusePs_ReadSecCtrlBits_Regs(
 	if (ReadOption == 0U) {
 		RegData = XilSKey_ReadReg(XSK_ZYNQMP_EFUSEPS_BASEADDR,
 				XSK_ZYNQMP_EFUSEPS_MISC_USER_CTRL_OFFSET);
-		Status = XST_SUCCESS;
+		Status = (u32)XST_SUCCESS;
 	}
 	else {
 		Status = XilSKey_ZynqMp_EfusePs_ReadRow(
